STSX  commit 10/17/2019
fabSuiteCom/fs.js
if passing XML to FS, needs lot number identified when they are in use
for the project, incl '' (blank/empty)
add <SHIP> to xml for fabsuite, was only <ShopFloorSave>
jobsts/globals.js
add rawprint screen that print raw label data
screenToggle onActionScreenToggle for window floating and sizing
when specifying toString() due to multi db make sure the variable is set
on Checklist Receive show message indicating zero remaining raw material
jobsts/jobs.js
importJob variable deleteByGuid, sheetImportOnly to id import at sheet
level for FS deleted guids to be done
guidsFabSuite [] track FS guids
currentGuidsPcmk [] current import GUIDs called from import_guids table
currentGuidsPcmkDB [] FS guids currenlty in STSX
currentGuidsPcmkParent [] parent of guid at same indic
createIdFildRecord
change create rec to bulk insert for speedup
importRecords_sheet()
deleteNonImportGuids() clear out guids no longer in FS
scopes.jobs.importJob.pcmksCreate - track raw sql data for piecemarks
scopes.jobs.importJob.idfilesCreate - track raw sql data for idfiles
currentPcmkInfo - track bcIds for all barcode numbers in use
	guidDistrib [] guidParent [] elsewhereID [] guids used elsewhere
flushAllClientsCache after raw sql import complete
saveSerials moved out of creation stream for raw SQL
indicate wrong label count during import if there are negative entries
initially and ignore by user
and reenable stopped import prep screen
correct pcmk join table pcmk_instances during import and clear out
unused entries
a lot of speedup changes to the code using caching variables and culling
data
add pcmk_instances table i18n to modify table 
clearUnique piecemarks in join table that should have been deleted
	not sure whether need to check other pcmk_instances that have moved
idfiles
rewrite arrays as entries into a raw SQL data and column statement
	sqlArgsToSqlData, sqlArrayToColumnNames
resequence and re-lot idfiles if changed for an incoming FS GUID
get simple piecemark different from STSX vs FS
getidnumUsedElsewhere() so the GUID no longer on a sequence and lot can
be identified
user printer and bartender templates were not showing correctly
show import selection view elements as DIRTY and in yellow if the id has
data
remove reprint,nohist,append,delete from import selections until covered
in discussion
	delete,nohist, reprint covered elsewhere. not sure about append
remove negative flag from sts_qty on import screen and show as dirty
process guids as raw sql due to large number of those guids during
import_table creation

+++ STS3/forms/kiss_import.js
@@ -257,6 +257,7 @@ function getKissFile(event){
 			<IncludeLotNumbers>1</IncludeLotNumbers>\n\
 			</ExportJob>\n\
 			</FabSuiteXMLRequest>';
+		//application.output(request);
 		if (vSeqNumber && vSeqNumber.search(',') == -1){
 			request = request.replace('SEQ','<Sequence>'+vSeqNumber+'</Sequence>');
 		} else {
@@ -319,6 +320,7 @@ function getKissFile(event){
 		application.output(file+' file write to '+servoyDir+ ' server success: '+success);
 		return;
 	}
+	scopes.jobs.importJob.sheetImportOnly = (!vSeqNumber && !vLotNumber && !vPartNumber);//can only delete items when sheet or full import only, use during import to STS db
 	removeJobImportData(vJobNumber);
 	elements.numSeq.requestFocus();
 	scopes.kiss.importFSOnServer(event,request,filters);//IMPORT 1 importFSOnServer
@@ -554,10 +556,11 @@ function removeJobImportData(jobNumber){
 	if (Q.getSize() > 0){
 		var rec = Q.getRecord(1);
 		jobId = rec.job_uuid;
+		Q.deleteAllRecords();
 	}
 	if (application.isInDeveloper()){application.output('size of import table records '+Q.getSize())}
 	//application.output('size of import table records '+Q.getSize())
-	Q.deleteAllRecords();
+
 	/** var tableIds = [];
 	/** @type {QBSelect<db:/stsservoy/import_table>} * /
 	var rec = null; var idx = 1;
@@ -576,12 +579,17 @@ function removeJobImportData(jobNumber){
 	var S = databaseManager.getFoundSet(s);
 	if (application.isInDeveloper()){application.output('size of import guid records '+S.getSize())}
 	//application.output('size of import guid records '+S.getSize())
-	success = S.deleteAllRecords();
+	if (S.getSize() > 0){
+		databaseManager.saveData(S);
+		success = S.deleteAllRecords();
+	}
 	if (application.isInDeveloper()){application.output('Delete job GUIDs: '+success)}
-	var success = Q.deleteAllRecords();
+	if (Q.getSize() > 0){
+		var success = Q.deleteAllRecords();
+	}
 	if (application.isInDeveloper()){application.output('Delete job import Records: '+success)}
 	null;
-	
+
 }
 /**
  * Handle hide window.
